Update README.md
Hi everyone , welcome to Powergentek. Today we will build an RFID-based solenoid door lock system using an Arduino board . The system uses an RFID RC522 reader module to read RFID cards , and a solenoid door lock to lock or unlock the door .
****************************************************************
The program allows for multiple RFID cards to be added or removed from the system, and a master card is used to perform administrative tasks such as adding or removing cards. This feature provides flexibility and convenience, as users can add or remove cards as needed without reprogramming the entire system.

The project includes an EEPROM module that can store up to 20 card data, allowing users to customize access to their homes or businesses. So this EEPROM based Arduino RFID door lock is a versatile project that offers a wide range of useful features for home security systems.